change to test lda callback update graph function
